bankingApi/requirements.txt
django==3.2.10
django-rest-framework==0.1.0
django-dotenv==1.4.2